requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python
numpy
pillow
pygame
